move app protect waf funtions to their own file
